refactored the layout of files
