+++ src/main/java/com/loot/server/domain/cards/PlayedCard.java
@@ -1,11 +1,10 @@
-package com.loot.server.socket.logic.cards.impl;
+package com.loot.server.domain.cards;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
