+++ src/main/java/com/loot/server/domain/request/PlayCardRequest.java
@@ -1,10 +1,9 @@
-package com.loot.server.domain;
+package com.loot.server.domain.request;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.loot.server.domain.dto.PlayerDto;
-import com.loot.server.socket.logic.cards.Card;
-import com.loot.server.socket.logic.cards.impl.PlayedCard;
+import com.loot.server.domain.cards.PlayedCard;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
