+++ src/main/java/com/loot/server/domain/response/Player.java
@@ -1,14 +1,13 @@
-package com.loot.server.domain;
+package com.loot.server.domain.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class GameCreationDto {
-    private String roomKey;
+public class Player {
+
+    String name;
 }
