+++ src/main/java/com/loot/server/service/GameService.java
@@ -1,7 +1,5 @@
 package com.loot.server.service;
 
-import com.loot.server.domain.GameCreationDto;
-
 public interface GameService {
     
     String generateRoomKey();
