--- src/main/java/com/loot/server/socket/logic/cards/ICardStack.java
@@ -1,11 +1,0 @@
-package com.loot.server.socket.logic.cards;
-
-public interface ICardStack {
-
-    void shuffle();
-
-    Integer drawCard();
-
-    Boolean deckIsEmpty();
-
-}
