+++ src/main/java/com/loot/server/socket/logic/impl/CardStack.java
@@ -1,6 +1,6 @@
-package com.loot.server.socket.logic.cards;
+package com.loot.server.socket.logic.impl;
 
-import com.loot.server.socket.logic.cards.impl.*;
+import com.loot.server.socket.logic.ICardStack;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
